starts writing tests for modThree
